+++ data/monitoring_coordinate.data
@@ -1,4 +1,4 @@
-Geoname ID;Name;ASCII Name;Coordinates
+Geoname_ID;Name;ASCII_Name;Coordinates
 2522696;Zafferana Etnea;Zafferana Etnea;37.67895, 15.10432
 2522749;Villamar;Villamar;39.61884, 8.95877
 2522847;Tropea;Tropea;38.67449, 15.89505
