+++ src/main/java/uni/climatemonitor/generics/Constants.java
@@ -1,0 +1,16 @@
+package uni.climatemonitor.generics;
+
+public class Constants {
+    /* some generic constants for graphics */
+    public static final String APP_NAME_S = "ClimaScope";
+    public static final String LOGO_PATH_S = "./src/main/images/logo.png";
+    public static final String EARTH_LOGO_PATH_S = "./src/main/images/earth_logo.png";
+
+    /* data paths */
+    public static final String CLIMATE_PARAMS_S = "./data/climate_parameters.data";
+    public static final String MONITORING_CENTERS_S = "./data/monitoring_centers.data";
+    public static final String MONITORING_COORDS_S = "./data/monitoring_coordinates.data";
+    public static final String REGISTERED_OPS_S = "./data/registered_operators.data";
+
+    public static final String[] prova = {"ciao", "come", "va", "io", "bene", "e", "tu", "?"};
+}
