--- src/main/java/uni/climatemonitor/graphics/Constants.java
@@ -1,8 +1,0 @@
-package uni.climatemonitor.graphics;
-
-public class Constants {
-    public static final String APP_NAME_S = "ClimaScope";
-    public static final String LOGO_PATH_S = "./src/main/images/logo.png";
-    public static final String EARTH_LOGO_PATH_S = "./src/main/images/earth_logo.png";
-    public static final String[] prova = {"ciao", "come", "va", "io", "bene", "e", "tu", "?"};
-}
